+++ web/src/pages/Logs/layout.css
@@ -552,7 +552,7 @@ button.logs-sidebar-backdrop.is-open {
   border: 1px solid var(--border);
   border-radius: 12px;
   padding: 1.25rem;
-  width: min(480px, 100%);
+  width: min(56rem, calc(100vw - 1.5rem));
   max-height: calc(100dvh - 2rem);
   overflow-y: auto;
   box-shadow: var(--shadow-md);
@@ -601,25 +601,27 @@ button.logs-sidebar-backdrop.is-open {
   margin-bottom: 0.75rem;
 }
 
-.logs-detail-image-wrap {
+.logs-detail-image-button {
+  display: block;
+  width: 100%;
+  border: 0;
+  border-radius: 8px;
+  padding: 0;
   margin-bottom: 0.75rem;
-  border-radius: 6px;
+  cursor: zoom-in;
   overflow: hidden;
-  max-height: 60dvh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
   background: var(--bg-subtle);
+  text-align: center;
 }
 
-.logs-detail-image-wrap .logs-thumb-button,
-.logs-detail-image-wrap .logs-thumb-image {
+.logs-detail-image {
+  display: block;
   width: auto;
   height: auto;
   max-width: 100%;
-  max-height: 60dvh;
+  max-height: 65dvh;
   object-fit: contain;
-  display: block;
+  margin: 0 auto;
 }
 
 .logs-detail-meta {
